refactor(cli): flatten feishu command module
Move the Feishu command domain to the neutral command root and update its internal exports without changing public CLI or API behavior.
