+++ turbo/apps/cli/src/commands/feishu/message/index.ts
@@ -2,7 +2,7 @@ import { Command } from "commander";
 
 import { sendCommand } from "./send";
 
-export const zeroFeishuMessageCommand = new Command()
+export const feishuMessageCommand = new Command()
   .name("message")
   .description("Send Feishu messages")
   .addCommand(sendCommand);
